// frontend/src/app/models/sanpham.model.ts
export class SanPham {
    id: number = 0;
    ten: string = "";
    moTa: string = "";
    anh: string = "";
    ngayTao: Date = new Date();
    trangThai: number = 0;
    idNhaSanXuat: number = 0;
    idLoai: number = 0;
    gia: number = 0;
    phanTram: number = 0;
    soLuong: number = 0;
    giaGiamGia: number = 0;
    tenNhaSanXuat: string = "";
    tenLoai: string = "";
}
  